+++ src/routes/+layout.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import Footer from '$lib/components/footer.svelte';
-  import Header from '$lib/components/header.svelte';
+  import Header from '$lib/components/header_.svelte';
   import {
     metaDescription,
     googleSiteVerification,
